fix: universe with Vec2 works now; kinematics returns copies of Vec2's
SpaceObjectModel's kinematics instance was returning references to the
position Vec2. The simulation of future paths would be run through, but
at each time step, a reference to the same position would be stored. The
position at that point in time wasn't getting captured. I fixed this by
having the kinematics getPosition() etc. methods return a copy of the
Vec2.